// client/src/generated/api/models/SectionRequestDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Event Ticketing System API
 * REST API for event ticketing platform with Stripe payment integration, JWT authentication, and QR code ticket generation
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SectionRequestDTO
 */
export interface SectionRequestDTO {
    /**
     * 
     * @type {string}
     * @memberof SectionRequestDTO
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof SectionRequestDTO
     */
    price: number;
    /**
     * 
     * @type {number}
     * @memberof SectionRequestDTO
     */
    rows: number;
    /**
     * 
     * @type {number}
     * @memberof SectionRequestDTO
     */
    cols: number;
}

/**
 * Check if a given object implements the SectionRequestDTO interface.
 */
export function instanceOfSectionRequestDTO(value: object): value is SectionRequestDTO {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('price' in value) || value['price'] === undefined) return false;
    if (!('rows' in value) || value['rows'] === undefined) return false;
    if (!('cols' in value) || value['cols'] === undefined) return false;
    return true;
}

export function SectionRequestDTOFromJSON(json: any): SectionRequestDTO {
    return SectionRequestDTOFromJSONTyped(json, false);
}

export function SectionRequestDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): SectionRequestDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'price': json['price'],
        'rows': json['rows'],
        'cols': json['cols'],
    };
}

export function SectionRequestDTOToJSON(json: any): SectionRequestDTO {
    return SectionRequestDTOToJSONTyped(json, false);
}

export function SectionRequestDTOToJSONTyped(value?: SectionRequestDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'price': value['price'],
        'rows': value['rows'],
        'cols': value['cols'],
    };
}